requirements.txt
google.generativeai
streamlit_chat
streamlit
python-dotenv
Pillow
google-cloud